refactor: add index root page and rename PrivateRouter
